fix: Phase 1 実機動作確認で発見したバグを修正
- omoikane_intent.py: attention_mask を .float() → .to(hidden_states.dtype) に修正
  （4bit量子化時の BFloat16/Float32 dtype mismatch を解消）
- tenson_korin_quantizer.py: protected_modules から lm_head を除外
  （lm_head を float16 に変換すると backbone の bfloat16 出力と不一致になるため）
- eval_baseline.py: llm-jp-eval v1.4.1 対応
  - Hydra config 形式への移行（サブコンフィグ分割、全必須キー追加）
  - データセットパスのバージョン付きサブディレクトリ自動検出
  - run_eval で絶対パスを使用し二重参照バグを修正
- .gitignore: external/, results/, .claude/ を追加

+++ kojiki_lm/iwato/omoikane_intent.py
@@ -115,7 +115,7 @@ class OmoikaneIntentRouter(nn.Module):
             return hidden_states.mean(dim=1)
 
         # マスクを拡張して d_model 次元に適用
-        mask = attention_mask.unsqueeze(-1).float()  # [batch, seq_len, 1]
+        mask = attention_mask.unsqueeze(-1).to(hidden_states.dtype)  # [batch, seq_len, 1]
         masked_hidden = hidden_states * mask
         sum_hidden = masked_hidden.sum(dim=1)          # [batch, d_model]
         count = mask.sum(dim=1).clamp(min=1.0)         # [batch, 1]
